IGNITE-22097 JobContext is not closed when job execution fails
